requirements.txt
streamlit
torch
opencv-python
numpy
matplotlib
Pillow
